+++ java/org/ohdsi/sql/TestSqlRender.java
@@ -1,24 +1,11 @@
 package org.ohdsi.sql;
 
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
 public class TestSqlRender {
 
 	public static void main(String[] args) {
-		String sql = "WITH cte_all\r\n" + 
-				"AS (\r\n" + 
-				"	SELECT * FROM @cdmSchema.MEASUREMENT m\r\n" + 
-				"		\r\n" + 
-				"	UNION ALL\r\n" + 
-				"	\r\n" + 
-				"	SELECT 	'(Glucose [Mass/volume] in Serum or Plasma --12 hours fasting)' AS check_description\r\n" + 
-				"	)\r\n" + 
-				"INSERT INTO @resultsSchema.@resultsTable\r\n" + 
-				"SELECT *\r\n" + 
-				"FROM cte_all;";
+		String sql = "DROP TABLE IF EXISTS #my_temp;";
 		String path = "inst/csv/replacementPatterns.csv";
-		sql = SqlTranslate.translateSqlWithPath(sql, "postgresql", null, null, path);
+		sql = SqlTranslate.translateSqlWithPath(sql, "sql server", null, null, path);
 		System.out.println(sql);
 		
 //		Pattern pattern = Pattern.compile("^((?!FROM).)*$");